+++ src/components/badge/index.ts
@@ -1,0 +1,19 @@
+import {
+  spectreBadgeVariants,
+  type SpectreBadgeVariant,
+  spectreInputSizes,
+  type SpectreInputSize,
+} from '../../utils/form';
+
+export {
+  defineSpectreBadge,
+  SpectreBadgeElement,
+  type SpectreBadgeProps,
+} from './sp-badge';
+
+export {
+  spectreBadgeVariants,
+  type SpectreBadgeVariant,
+  spectreInputSizes as spectreBadgeSizes,
+  type SpectreInputSize as SpectreBadgeSize,
+};