+++ src/components/icon-box/index.ts
@@ -1,0 +1,19 @@
+import {
+  spectreIconBoxVariants,
+  type SpectreIconBoxVariant,
+  spectreInputSizes,
+  type SpectreInputSize,
+} from '../../utils/form';
+
+export {
+  defineSpectreIconBox,
+  SpectreIconBoxElement,
+  type SpectreIconBoxProps,
+} from './sp-icon-box';
+
+export {
+  spectreIconBoxVariants,
+  type SpectreIconBoxVariant,
+  spectreInputSizes as spectreIconBoxSizes,
+  type SpectreInputSize as SpectreIconBoxSize,
+};